import style from file
