+++ src/lib/components/SearchBar/searchbar.tsx
@@ -1,7 +1,7 @@
 import { ChangeEvent } from 'react';
 import { TextField, InputAdornment, IconButton } from '@mui/material';
 import SearchIcon from '@mui/icons-material/Search';
-import { useTheme } from '@mui/material/styles';
+import { useCustomStyles } from './style';
 
 interface SearchBarProps {
   onSearch: (term: string) => void;
@@ -11,53 +11,8 @@ const SearchBar: React.FC<SearchBarProps> = ({ onSearch }) => {
   const handleSearch = (event: ChangeEvent<HTMLInputElement>) => {
     onSearch(event.target.value);
   };
-  const theme = useTheme();
 
-  const textFieldStyle = {
-    '& .MuiInputBase-root': {
-      backgroundColor: 'rgba(229, 255, 233, 0.1)',
-      borderRadius: '20px',
-      borderColor: theme.customPalette.primary.main,
-      color: theme.customPalette.primary.background,
-      typography: theme.customTypography.h7,
-      fontFamily: theme.customTypography.fontFamily,
-      '&:hover fieldset': {
-        borderColor: theme.customPalette.secondary.main,
-      },
-      '& fieldset': {
-        border: '2px solid',
-        borderColor: theme.customPalette.primary.main,
-      },
-      '&.Mui-focused fieldset': {
-        borderColor: theme.customPalette.secondary.main,
-      },
-    },
-    '& .MuiInputBase-input': {
-      '&::placeholder': {
-        opacity: 1,
-        color: theme.customPalette.neutral.inactive,
-        typography: theme.customTypography.h7,
-      },
-    },
-  };
-
-  const InputStyle = {
-    '& .MuiButtonBase-root': {
-      border: '2px solid',
-      borderColor: theme.customPalette.primary.main,
-      backgroundColor: theme.customPalette.primary.main,
-      '&:hover': {
-        backgroundColor: theme.customPalette.secondary.main,
-      },
-    },
-  };
-
-  const SearchIconStyle = {
-    color: theme.customPalette.primary.background,
-    '&hover': {
-      color: theme.customPalette.primary.background,
-    },
-  };
+  const style = useCustomStyles();
 
   return (
     <TextField
@@ -65,12 +20,12 @@ const SearchBar: React.FC<SearchBarProps> = ({ onSearch }) => {
       fullWidth
       placeholder="Busque o seu produto..."
       onChange={handleSearch}
-      sx={textFieldStyle}
+      sx={style.textFieldStyle}
       InputProps={{
         endAdornment: (
-          <InputAdornment position="end" sx={InputStyle}>
+          <InputAdornment position="end" sx={style.inputStyle}>
             <IconButton>
-              <SearchIcon sx={SearchIconStyle} />
+              <SearchIcon sx={style.searchIconStyle} />
             </IconButton>
           </InputAdornment>
         ),
