--- src/lib/components/SearchBar/styles.js
@@ -1,1 +1,0 @@
-
